refactor: replace to next-mdx-remote
