+++ lib/share.ts
