+++ package.json
@@ -5,12 +5,10 @@
     "@prisma/client": "^3.1.1",
     "date-fns": "^2.25.0",
     "next": "^11.1.2",
+    "next-mdx-remote": "^3.0.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "remark": "^14.0.1",
-    "remark-gfm": "^3.0.0",
-    "remark-html": "^15.0.0",
     "styled-components": "^5.3.1"
   },
   "description": "Shindan by AniMare",
