+++ pages/s/[slug].tsx
@@ -9,8 +9,8 @@ import Layout from '../../components/layout'
 import MessageWindow from '../../components/message-window'
 import SEO from '../../components/seo'
 import { getAnalysisResultIDs, getAnalysisResult } from '../../lib/analysis'
+import { createTweetURL } from '../../lib/share'
 import pkg from '../../package.json'
-import { createTweetURL } from '../../utils/share'
 
 import { GetStaticPaths, GetStaticProps, NextPage } from 'next'
 import type { Result } from '../../lib/analysis'
